Move library transformations to their own package
