+++ src/main/scala/com/leobenkel/safetyplugin/Transformations/MergeStrategyConfiguration.scala
@@ -1,4 +1,4 @@
-package com.leobenkel.safetyplugin
+package com.leobenkel.safetyplugin.Transformations
 
 import sbtassembly.{MergeStrategy, PathList}
